in the encrypDecrypt.c and .h files, it is currently uint16, and it can handle anything less than 255 elements, but if its above it dies

+++ library/encryptDecrypt.c
@@ -5,7 +5,7 @@
 void encryptHMAC(uint8_t* output, uint16_t* msgLength, uint8_t* data, uint8_t* inputKey, uint16_t keyLength){
 	uint8_t key[encryptKeyLength + macKeyLength] = {0};
 	uint8_t macKey_String[macKeyLength*2] = {0};
-	uint16_t IV[IVlength] = {0};
+	uint8_t IV[IVlength] = {0};
 	
 	uint8_t IV_ciphertextLength = 0;
 	uint8_t IVciphertextConcat[IVlength + MAX_TRANSMISSION_BLOCK_LENGTH + 16] = {0}; /* 16 is a safety number */
@@ -87,7 +87,7 @@ void decryptHMAC(uint8_t* output, uint16_t* msgLength, uint8_t* registKey, uint8
 }
 
 
-void encrypt(uint8_t* output, uint16_t* msgLength, uint8_t* data, uint8_t* inputKey, uint16_t keyLength){
+void encrypt(uint8_t* output, uint8_t* msgLength, uint8_t* data, uint8_t* inputKey, uint16_t keyLength){
 	uint8_t key[encryptKeyLength + macKeyLength] = {0};
 	uint8_t macKey_String[macKeyLength*2] = {0};
 	uint8_t IV[IVlength] = {0};
@@ -130,7 +130,7 @@ void encrypt(uint8_t* output, uint16_t* msgLength, uint8_t* data, uint8_t* input
  }
 
 
-void decrypt(uint8_t* output, uint16_t* msgLength, uint8_t* registKey, uint8_t* inputKey, uint16_t keyLength){
+void decrypt(uint8_t* output, uint8_t* msgLength, uint8_t* registKey, uint8_t* inputKey, uint16_t keyLength){
 	uint8_t key[encryptKeyLength + macKeyLength] = {0};	
 	uint8_t macKey_String[macKeyLength*2] = {0};
 	uint8_t registKey_String[(IVlength + MAX_TRANSMISSION_BLOCK_LENGTH + SHA256_DIGEST_LENGTH)*2] = {0};
@@ -195,6 +195,5 @@ Decryption
 =	 	  		output	 				...(3)
 
 * * * * * * * * * * * * * * * * * * * * */
-
